requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
djangorestframework==3.13.1
idna==3.3
lxml==4.8.0
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.9
django-wrapwith==0.0.2
django-mathfilters==1.0.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.2.1
jupyterlab-widgets==1.1.0
